chore(post): move package
