+++ src/main/java/com/example/cheerdo/post/service/PostServiceImpl.java
@@ -6,14 +6,12 @@ import com.example.cheerdo.entity.Post;
 import com.example.cheerdo.post.dto.request.LetterRequestDto;
 import com.example.cheerdo.post.dto.request.PostRequestDto;
 import com.example.cheerdo.post.dto.response.LetterResponseDto;
-import com.example.cheerdo.post.dto.response.PostResponseDto;
+import com.example.cheerdo.post.repository.MemberRepository;
 import com.example.cheerdo.post.repository.PostRepository;
 import com.example.cheerdo.post.repository.RelationRepository;
 import lombok.AllArgsConstructor;
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Service;
 
-import javax.management.relation.Relation;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.Optional;
